+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changes
 
 - Rewrite edit link on server to navigate to Tekir admin page.
 - Add support for adding attachments.
+- Add support for opening relevant folders in file manager.
 - Make UI language preference persistent.
 - Add preview button to contents page.